trying to get float

+++ assignment.py
@@ -1,6 +1,6 @@
 print("What is the diameter of the circle?")
-D = (input())
-radius = (D / 2)
+D = float(input())
+radius = float(D / 2)
 A = float((radius ** 2) * 3.14)
 C = float(2 * 3.14 * radius)
 print(float(A))
